Fixup integration handling
Installation_id is not a one off configurable option that can be stored.
Luckily it's available in webhook events that are sent through the
integration. Getting this passed through zuul and back as an event is
going to be difficult, however the same project always has the same
installation_id so if we just save this when a webhook is received we
can fetch it again later.

Change-Id: I917e09d79a4eca3172ee86321a437fe3fe0eb424
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>