Manually close pull request with integration
There's currently a bug in github's integration code that means that
even though the pull request is merged it's being left in an open state.
I've reported it at [1] and has been raised by others [2].

For now the best we can do is manually close the pull request, however
it's concerning that in [2] it mentions that it also doesn't not close
associated Issues, build pages, or other things that typically will
trigger on merge.

[1] https://platform.github.community/t/pull-requests-open-after-merge/1460
[2] https://platform.github.community/t/pushing-a-pull-request-doesnt-close-it/1112

Change-Id: I8c238d0972b496650ab1dc9fa69ae7d279fcf19c
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>