Don't lookup github projects not in layout.
Gerrit does a funny thing by fetching change information you receive -
even if you don't have that project in your layout. This is supposed to
update the change cache for any cross-project dependencies on foreign
projects. The project object is None that gets passed to getChange and
gerrit only uses event information to fetch data.

We can't do this with github yet, but what we get is a

    'NoneType' object has no attribute 'name'

message when you try to fetch this empty project. Probably we should be
fetching based on event.project_name but really we can't support it
anyway so just short circuit and return a NullChange.

Change-Id: Ib03c428e7fe82bd62530f0a8fc834dae6648fd58
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>