Quote ansible command and include config file
One of the useful parts of the executor debug output is that if you have
keep-jobdir flag on and are holding the node you can run the command
again to debug any problems.

Unfortunately you have to include the ANSIBLE_CONFIG file as an env
variable to make this work - and there is no CLI equivalent for
ANSIBLE_CONFIG. Include the config file in the debug output and unquote
the cmd for easier copy and pasting.

Change-Id: If968f8f0c20f0d043653e8c127603b1725875d89