If installation_key is not found revert to API key
Hacks on hacks. We already maintain a map of projects that are using an
integration so that we can fetch the installation id. If the project is
not in this list it's probably because they are not using an integration
and so we should fall back to using an api key.

Change-Id: I1fcfb75c93de0a34e10eead6f713087ea41d3586
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>