A docker-compose solution for running zuul tests
Instead of having to customize zookeeper and mysql on your machine to
run zuul tests, lets bring everything up together in a series of
containers and run the tests that way.

As of writing this is super slow and has strange thread errors. No idea
why. Other people have found similar results trying to run in docker.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>